Address pr comments: Change generateKey back from utils to controller_framework.go
